Update script.js
maybe fixed the code, used chatgpt to check it because I wanted to try it, never used it for code before (not that I use it for anything else). I had a few mistakes, some spots needed to be myCanvas instead of ctx, and I don't need the onload to get the setInterval to run. also moved the beginpath to the function and added ctx.fill to it.